fix: use relative URL in rosdoc directive

Closes SN-870

See merge request robot/ned/ned-ros-stack!238

+++ niryo_robot_utils/src/niryo_robot_utils/dataclasses/JointsPosition.py
@@ -40,7 +40,7 @@ class JointsPosition(ABCSerializable):
     def __eq__(self, other) -> bool:
         return self.__joints == other.__joints
 
-    def to_list(self) -> List[int]:
+    def to_list(self) -> List[float]:
         return list(self.__joints)
 
     def to_dict(self) -> Dict[str, Any]:
